requirements/base.txt
Django==2.2.2
pillow==6.0.0
django-debug-toolbar==1.11
djangorestframework==3.9.4
djangorestframework-xml==1.4.0
django-filter==2.1.0
requests==2.22.0
psycopg2-binary==2.8.2
coverage==4.5.3
mixer==6.1.3
